+++ src/orientation.rs
@@ -1,7 +1,7 @@
 /*
     src/orientation.rs
 
-    Copyright (c) 2019 Stephen Whittle  All rights reserved.
+    Copyright (c) 2019, 2020 Stephen Whittle  All rights reserved.
 
     Permission is hereby granted, free of charge, to any person obtaining a copy
     of this software and associated documentation files (the "Software"),
@@ -27,4 +27,13 @@ pub enum Orientation {
     Top,
     /// Rip a line from the bottom of the screen.
     Bottom
+}
+
+impl Orientation {
+    pub(in crate) fn value(self) -> i32 {
+        match self {
+            Orientation::Top    => 1,
+            Orientation::Bottom => -1
+        }
+    }
 }