+++ CHANGELOG.md
@@ -9,8 +9,8 @@ All breaking changes are marked with [BC] and potentially require API consumer c
 - `ncursesw::menu::menu_driver()` now returns a `Result<Option<MenuRequest>, NCurseswMenuError>` instead of `Result<Option<i32>, NCurseswMenuError>`. [BC]
 - `bindgen` will now only create bindings against version 6.1 and above of the NCurses library.
 - `ncursesw::newterm()` implemented (was calling `unimplemented!()` and signature now takes `O: std::os::unix::io::AsRawFd + std::io::Write` and `I: std::os::unix::io::AsRawFd + std::io::Write` instead of `shims::bindings::FILE` for both. [BC]
-- `ncursesw::getwin()` now takes `I: std::os::unix::io::AsRawFD + std::io::Read` instead of `&std::path::Path`. [BC]
-- `ncursesw::putwin()` now takes `O: std::os::unix::io::AsRawFD + std::io::Write` instead of `&std::path::Path`. [BC]
+- `ncursesw::getwin()` now takes `I: std::os::unix::io::AsRawFd + std::io::Read` instead of `&std::path::Path`. [BC]
+- `ncursesw::putwin()` now takes `O: std::os::unix::io::AsRawFd + std::io::Write` instead of `&std::path::Path`. [BC]
 - `ncursesw::shims::ncurses::wunctrl()` now returns `Option<*mut wchar_t>` instead of `*mut wchar_t`. [BC]
 - Added missing NCurses `_sp` (screen) functions.